// src/modules/orders/types/order.types.ts
export interface CreateOrderRequest {
    catalog_id: string;
    items: {
        product_id: string;
        quantity: number;
    }[];
}

export interface OrderSummary {
    id: string;
    total_items: number;
    created_at: string;
}

export interface OrderItem {
    product_id: string;
    product_name: string;
    price: number;
    quantity: number;
}

export interface OrderDetail {
    id: string;
    total_items: number;
    created_at: string;
    order_items: OrderItem[];
}

export interface PlaceOrderResponse {
    whatsapp_url: string;
}